EU flag
Source:
https://commons.wikimedia.org/wiki/File:Flag_of_Europe.svg

Public Domain, but:

The Council of Europe holds the copyright for the European flag.

Requests shall be addressed to the Legal Affairs Directorate of the
Council of Europe: [F-67075 Strasbourg Cedex; tel. +33-388 41 20 00; fax
+33-388 41 20 52]

The European emblem may be used only if:

there is no likelihood of the user of the emblem being confused with the
European Community or the Council of Europe;
the emblem is not used in connection with objectives or activities which
are incompatible with the aims and principles of the European Community
or of the Council of Europe.
Permission to use the European emblem does not confer on those to whom
it is granted any right of exclusive use, nor does it allow them to
appropriate the emblem or any similar trademark or logo, either by
registration or any other means. Each case will be examined individually
to ascertain whether it satisfies the criteria set out above. This will
be unlikely in a commercial context if the European emblem is used in
conjunction with a company's own logo, name or trade mark.