Look up any git commit
In preparation for #167 so we can build raw URIs (Really the datamodel
is upside-down to support that at the moment)

Should probably return a Stream rather than a List for effiency (later
we'll just pick the first one that has the commit).

also.. is it possible to return the RetrievedFrom object instead of
Workflow?