Fix MongoDB test
Had to add a new dependency for the embedded MongoDB server used for tests.

The version of MongoDB used for tests (embedded) also needs to be defined now. And
MongoDB/SpringData integration appears to be more strict. A query like
{ obj.property: ?0 } failed, and took me a long time to recognize that it was missing
quotes, so { 'obj.property': ?0 } fixed it.

Finally, there was also an issue with serialization. For some reason now the Mongo
repository was failing to serialize a response for Workflow. The object contained a
final String that was used for the base URL of permanent links, and Mongo/Spring/etc
wanted a way to set that value in the constructor. Tried the @PersistenceConstructor
annotation but that didn't work. In the end it was easy to fix it by making the
final String now a constant